requirements.txt
numpy==1.24.2
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tqdm==4.65.0
tzdata==2023.3
